# Avoid recompiling every SQL query