--- a/src/Microsoft.Health.Fhir.SqlServer/Features/Search/Expressions/Visitors/QueryGenerators/QuantityQueryGenerator.cs
+++ b/src/Microsoft.Health.Fhir.SqlServer/Features/Search/Expressions/Visitors/QueryGenerators/QuantityQueryGenerator.cs
@@ -57,7 +57,7 @@ namespace Microsoft.Health.Fhir.SqlServer.Features.Search.Expressions.Visitors.Q
                         .Append(" WHERE ")
                         .Append(VLatest.QuantityCode.Value, null)
                         .Append(" = ")
-                        .Append(context.Parameters.AddParameter(VLatest.QuantityCode.Value, expression.Value))
+                        .Append(context.Parameters.AddParameter(VLatest.QuantityCode.Value, expression.Value, true))
                         .Append(")");
 
                     return context;
@@ -75,7 +75,7 @@ namespace Microsoft.Health.Fhir.SqlServer.Features.Search.Expressions.Visitors.Q
                         .Append(" WHERE ")
                         .Append(VLatest.System.Value, null)
                         .Append(" = ")
-                        .Append(context.Parameters.AddParameter(VLatest.System.Value, expression.Value))
+                        .Append(context.Parameters.AddParameter(VLatest.System.Value, expression.Value, true))
                         .Append(")");
 
                     return context;