… src/Microsoft.Health.Fhir.SqlServer/Features/Search/SqlServerSearchService.cs