## Description

We currently add `OPTION(RECOMPILE)` to every query we generate for searches. The reason for this is that the optimizer chooses different query plans based on parameter values and sometimes there can be an enormous differences in performance between plans.

But there is a cost to recompiling queries:

|                      | Logical reads WITHOUT caching | Duration WITHOUT caching (ms) | Logical reads WITH cache hit| Duration WITH cache hit (ms) |
|----------------------|-------------------------------|-------------------------------|----------------------------|---------------------------|
| /Patient?gender=male | 191                           | 15                            | 39                         | 2                         |
| /Patient?family=Smith&birthdate=1982-11-05| 631                           | 57                            | 21                         | 3                         |
| /MedicationRequest?subject:Patient.identifier=999-60-8644 | 382                           | 32                            | 3                          | 4                         |

This PR makes a first change in avoiding recompilations for every query. We remove the `OPTION(RECOMPILE)` hint, and instead add a hash of a subset of the parameter values in a comment. This means that two identical searches will have the same query text and can use the same query plan. 

Excluded from the hashes are `TOP` values and parameters related to continuation tokens (so we don't require recompiling for every page of a query).

We'll want to to more here; this is just a first step. A promising option is 
[`OPTIMIZE FOR ( @variable_name { UNKNOWN | = <literal_constant> } [ , ...n ] ) ` ](https://docs.microsoft.com/en-us/sql/t-sql/queries/hints-transact-sql-query?view=sql-server-ver15).

## Testing
Manual and unit tests